requirements.txt
Package            Version
------------------ -----------
attrs              25.1.0
beautifulsoup4     4.13.3
certifi            2025.1.31
cffi               1.17.1
charset-normalizer 3.4.1
h11                0.14.0
idna               3.10
outcome            1.3.0.post0
pip                25.0.1
pycparser          2.22
PySocks            1.7.1
python-dotenv      1.0.1
requests           2.32.3
selenium           4.28.1
sniffio            1.3.1
sortedcontainers   2.4.0
soupsieve          2.6
trio               0.29.0
trio-websocket     0.12.1
typing_extensions  4.12.2
urllib3            2.3.0
websocket-client   1.8.0
wsproto            1.2.0
